First draft of an Actual Deposit
Moves Item related features to own Controller as I created routes for
viewing mets, downloading package and sending package which will be
quite useful for this stage of development even if they are removed
later (or only available to analyst or whatever).

Should be moved to a worker as it is a blocking request in this sample.
