+++ app/controllers/Application.scala
@@ -26,7 +26,6 @@ import services.contentModelJson._
 import services.publisherModelJson._
 import services.subscriberModelJson._
 import services.Emailer
-import workers.Packager
 
 case class HubContext(user: Option[User])
 
@@ -75,42 +74,6 @@ object Application extends Controller {
         Redirect(routes.Application.index)
       }
     )
-  }
-
-  def item(id: Int) = Action { implicit request =>
-    Item.findById(id).map( item =>
-      Ok(views.html.item.show(item, Subscriber.findByUserId(1)))
-    ).getOrElse(NotFound(views.html.static.trouble("No such item: " + id)))
-  }
-
-  def itemBrowse(filter: String, id: Int, page: Int) = Action { implicit request =>
-    filter match {
-      case "collection" => itemBrowseCollection(id, page)
-      case "topic" => itemBrowseTopic(id, page)
-      case _ => NotFound(views.html.static.trouble("No such filter: " + filter))
-    }
-  }
-
-  private def itemBrowseTopic(id: Int, page: Int)(implicit request: Request[AnyContent]): Result = {
-    Topic.findById(id).map( topic =>
-      Ok(views.html.item.browse(id, topic.pagedItems(page, 10), "topic", topic.name, page, topic.itemCount))
-    ).getOrElse(NotFound(views.html.static.trouble("No such topic: " + id)))
-  }
-
-  private def itemBrowseCollection(id: Int, page: Int)(implicit request: Request[AnyContent]): Result = {
-    Collection.findById(id).map( coll =>
-      Ok(views.html.item.browse(id, Item.inCollection(id, page), "collection", coll.description, page, Item.collectionCount(coll.id)))
-    ).getOrElse(NotFound(views.html.static.trouble("No such collection")))
-  }
-
-  def itemPackage(id: Int) = Action { implicit request =>
-    Item.findById(id).map( item =>
-      //Ok(item.toMets)
-      Result(
-        header = ResponseHeader(200, Map(CONTENT_TYPE -> "application/zip")),
-        body = Enumerator.fromStream(Packager.packageItem(item))
-      )
-    ).getOrElse(NotFound(views.html.static.trouble("No such item: " + id)))
   }
 
   def topics = Action { implicit request =>
