+++ app/views/topic/show.scala.html
@@ -26,10 +26,10 @@
          }
        }
     </div>
-    <h3>Recent Articles <a href="@routes.Application.itemBrowse("topic", topic.id, 0)">Browse All</a></h3>
+    <h3>Recent Articles <a href="@routes.ItemController.itemBrowse("topic", topic.id, 0)">Browse All</a></h3>
     <dl>
      	@topic.recentItems(6).map { i =>
-     	  <dt><a href="@routes.Application.item(i.id)">@Html(i.metadataValue("title"))</a></dt>
+     	  <dt><a href="@routes.ItemController.item(i.id)">@Html(i.metadataValue("title"))</a></dt>
         <dd>@tags.item_authors(i.metadataValues("author"))</dd>
       }
     </dl>
