+++ conf/routes
@@ -86,9 +86,11 @@ GET     /resmap/:id/scheme          controllers.Application.removeResourceMappin
 # Item pages
 #GET     /items                       controllers.Application.items
 #GET     /items/search                controllers.Application.itemSearch(topicId: String, page: Int)
-GET     /items/browse                controllers.Application.itemBrowse(filter: String, id: Int, page: Int ?= 0)
-GET     /item/:id                    controllers.Application.item(id: Int)
-GET     /item/package/:id            controllers.Application.itemPackage(id: Int)
+GET     /items/browse                controllers.ItemController.itemBrowse(filter: String, id: Int, page: Int ?= 0)
+GET     /item/:id                    controllers.ItemController.item(id: Int)
+GET     /item/package/:id            controllers.ItemController.itemPackage(id: Int)
+GET     /item/deposit/:id            controllers.ItemController.itemDeposit(id: Int)
+GET     /item/mets/:id               controllers.ItemController.itemMets(id: Int)
 
 # Topic pages
 GET     /topics                      controllers.Application.topics
